feat: add Netlify deployment support
